# business/sys/database/models.go
// Code generated by sqlc. DO NOT EDIT.

package omslitedb

import ()

type Order struct {
	ID          int64  `json:"id"`
	Productname string `json:"productname"`
	Quantity    int64  `json:"quantity"`
}