# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.JSONResult:
    properties:
      status:
        type: string
    type: object
  omslitedb.CreateOrderParams:
    properties:
      productname:
        type: string
      quantity:
        type: integer
    type: object
  omslitedb.Order:
    properties:
      id:
        type: integer
      productname:
        type: string
      quantity:
        type: integer
    type: object
host: localhost:8080
info:
  contact: {}
  description: oms-lite API, you know
  license:
    name: Apache 2.0
  title: oms-lite API
  version: "1.0"
paths:
  /orders:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/omslitedb.Order'
            type: array
      summary: list all orders
      tags:
      - orders
  /orders/:
    post:
      consumes:
      - application/json
      parameters:
      - description: Create order with product&quantity
        in: body
        name: order
        required: true
        schema:
          $ref: '#/definitions/omslitedb.CreateOrderParams'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.JSONResult'
            - properties:
                orderID:
                  type: integer
              type: object
      summary: create new order
      tags:
      - orders
  /orders/{order_id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.JSONResult'
      summary: delete order by id
      tags:
      - orders
    get:
      consumes:
      - application/json
      parameters:
      - description: Order ID
        in: path
        name: order_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/handlers.JSONResult'
            - properties:
                order:
                  $ref: '#/definitions/omslitedb.Order'
              type: object
      summary: get order by id
      tags:
      - orders
swagger: "2.0"